requirements.txt
PythonForWindows
pycryptodome
